First optimization algorithm
I bring the first optimization algorithm for voxels.
I refactored the Voxel code, leaving only a little information to be saved.
I added the BlockType structure, whose objective is to identify block types, such as DEFAULT and AIR
Blocks that have their faces in contact with another AIR block or at the end of the edge will have that face rendered, otherwise the face will not be rendered.
This allowed us to gain extraordinary performance, but not yet enough to generate large maps.