My First GreedyMesh
This version contains my first GreedyMesh with the Culling off from the last update
Basically it boils down to first using Culling Off to find out which voxels need to be rendered and then GreedyMesh comes in looking for the same type of voxel and creating a new mesh to be rendered.
GreedyMesh is still very raw, so it is not as efficient and only works in the x direction.
Chunk has also undergone an update, now we render voxels, we only render meshes, which are generated by GreedyMesh.