Code rewritted and reorganized
In this new update, I completely rewrote the code, reducing the number of unnecessary items and now we have a clearer and more functional hierarchy. The main updates were that GameObjects will now be responsible for Updates and Mesh will be responsible for rendering. This makes it clearer what the responsibility of each object in the Engine is.