Code Rewritten and Greedy Meshing
After months stopped, I returned.
I rewrote the entire code, introducing a clearer structure of hierarchies and inheritance, which was also faster.
I separated the main functions such as Render and Update into 2 objects, Mesh and GameObject.
The GameObject will be responsible for the update functions and the Mesh for the rendering functions
I really like the Unity Engine object structure and have plans to do something similar.
The code was completely rewritten from scratch, taking advantage of some small parts, such as the Camera and the ImGUI implementation.
We also have new features, such as simple structures of 3 integers and 3 floats and also a Sleep system, in which it is possible to execute functions with a timeout without blocking the main Thread and crashing the game.
This commit also brings an important algorithm for the operation of any game, which is GreedyMesh, it is working 100%, but it still does not handle different colors and textures, but it is already capable of merging hundreds of meshes and optimizing performance.
In my tests I managed to reduce 7200 triangles to just 60, a huge performance gain.
It is implementing directly in Chunk, but I intend to leave it aside and be able to call it from anywhere it needs optimization.
It was a big challenge for me to implement this algorithm, but I managed to make it work, some variables will be in Portuguese as I made it easier for better understanding, but I intend to switch to English and add comments on what each variable does